% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bgzip_compress}
\alias{bgzip_compress}
\title{bgzip compress with checks}
\usage{
bgzip_compress(output_name, file)
}
\arguments{
\item{output_name}{file to be compressed}

\item{file}{location to be saved}
}
\description{
bgzip compress with checks
}
