% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFun_BIC_Meng.R
\name{CalcBIC}
\alias{CalcBIC}
\title{function for BIC calculation}
\usage{
CalcBIC(y, PC, K)
}
\arguments{
\item{y}{length N vector}

\item{PC}{matrix of principal components with N rows and P columns}

\item{K}{kinship matrix with N rows and N columns}
}
\description{
function for BIC calculation
}
