% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateMAF}
\alias{calculateMAF}
\title{Updated MAF function}
\usage{
calculateMAF(df, ploidy)
}
\arguments{
\item{df}{fill description}

\item{ploidy}{fill description}
}
\description{
Updated MAF function
}
