% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMplot.r
\name{CMplot_shiny}
\alias{CMplot_shiny}
\title{Describe functin}
\usage{
CMplot_shiny(
  Pmap,
  col = c("#4197d8", "#f8c120", "#413496", "#495226", "#d60b6f", "#e66519", "#d581b7",
    "#83d3ad", "#7c162c", "#26755d"),
  bin.size = 1e+06,
  bin.breaks = NULL,
  LOG10 = TRUE,
  pch = 19,
  type = "p",
  band = 1,
  H = 1.5,
  ylim = NULL,
  axis.cex = 1,
  axis.lwd = 1.5,
  lab.cex = 1.5,
  lab.font = 2,
  plot.type = c("m", "c", "q", "d"),
  multracks = FALSE,
  multracks.xaxis = FALSE,
  multraits = FALSE,
  points.alpha = 100L,
  r = 0.3,
  cex = c(0.5, 1, 1),
  outward = FALSE,
  ylab = expression(-log[10](italic(p))),
  ylab.pos = 3,
  xticks.pos = 1,
  mar = c(3, 6, 3, 3),
  mar.between = 0,
  threshold = NULL,
  threshold.col = "red",
  threshold.lwd = 1,
  threshold.lty = 2,
  amplify = TRUE,
  signal.cex = 1.5,
  signal.pch = 19,
  signal.col = NULL,
  signal.line = 2,
  highlight = NULL,
  highlight.cex = 1,
  highlight.pch = 19,
  highlight.type = "p",
  highlight.col = "red",
  highlight.text = NULL,
  highlight.text.col = "black",
  highlight.text.cex = 1,
  highlight.text.font = 3,
  chr.labels = NULL,
  chr.border = FALSE,
  chr.labels.angle = 0,
  chr.den.col = "black",
  chr.pos.max = FALSE,
  cir.band = 1,
  cir.chr = TRUE,
  cir.chr.h = 1.5,
  cir.axis = TRUE,
  cir.axis.col = "black",
  cir.axis.grid = TRUE,
  conf.int = TRUE,
  conf.int.col = NULL,
  file.output = TRUE,
  file.name = "",
  file = c("jpg", "pdf", "tiff", "png"),
  dpi = 300,
  height = NULL,
  width = NULL,
  main = NULL,
  main.cex = 1.5,
  main.font = 2,
  legend.ncol = NULL,
  legend.cex = 1,
  legend.pos = c("left", "middle", "right"),
  box = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Pmap}{describe documentation}

\item{col}{describe documentation}

\item{bin.size}{describe documentation}

\item{bin.breaks}{describe documentation}

\item{LOG10}{describe documentation}

\item{pch}{describe documentation}

\item{type}{describe documentation}

\item{band}{describe documentation}

\item{H}{describe documentation}

\item{ylim}{describe documentation}

\item{axis.cex}{describe documentation}

\item{axis.lwd}{describe documentation}

\item{lab.cex}{describe documentation}

\item{lab.font}{describe documentation}

\item{plot.type}{describe documentation}

\item{multracks}{describe documentation}

\item{multracks.xaxis}{describe documentation}

\item{multraits}{describe documentation}

\item{points.alpha}{describe documentation}

\item{r}{describe documentation}

\item{cex}{describe documentation}

\item{outward}{describe documentation}

\item{ylab}{describe documentation}

\item{ylab.pos}{describe documentation}

\item{xticks.pos}{describe documentation}

\item{mar}{describe documentation}

\item{mar.between}{describe documentation}

\item{threshold}{describe documentation}

\item{threshold.col}{describe documentation}

\item{threshold.lwd}{describe documentation}

\item{threshold.lty}{describe documentation}

\item{amplify}{describe documentation}

\item{signal.cex}{describe documentation}

\item{signal.pch}{describe documentation}

\item{signal.col}{describe documentation}

\item{signal.line}{describe documentation}

\item{highlight}{describe documentation}

\item{highlight.cex}{describe documentation}

\item{highlight.pch}{describe documentation}

\item{highlight.type}{describe documentation}

\item{highlight.col}{describe documentation}

\item{highlight.text}{describe documentation}

\item{highlight.text.col}{describe documentation}

\item{highlight.text.cex}{describe documentation}

\item{highlight.text.font}{describe documentation}

\item{chr.labels}{describe documentation}

\item{chr.border}{describe documentation}

\item{chr.labels.angle}{describe documentation}

\item{chr.den.col}{describe documentation}

\item{chr.pos.max}{describe documentation}

\item{cir.band}{describe documentation}

\item{cir.chr}{describe documentation}

\item{cir.chr.h}{describe documentation}

\item{cir.axis}{describe documentation}

\item{cir.axis.col}{describe documentation}

\item{cir.axis.grid}{describe documentation}

\item{conf.int}{describe documentation}

\item{conf.int.col}{describe documentation}

\item{file.output}{describe documentation}

\item{file.name}{describe documentation}

\item{file}{describe documentation}

\item{dpi}{describe documentation}

\item{height}{describe documentation}

\item{width}{describe documentation}

\item{main}{describe documentation}

\item{main.cex}{describe documentation}

\item{main.font}{describe documentation}

\item{legend.ncol}{describe documentation}

\item{legend.cex}{describe documentation}

\item{legend.pos}{describe documentation}

\item{box}{describe documentation}

\item{verbose}{describe documentation}
}
\description{
Describe functin
}
