% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findK}
\alias{findK}
\title{Internal function}
\usage{
findK(genotypeMatrix, maxK, ploidy)
}
\arguments{
\item{genotypeMatrix}{fill description}

\item{maxK}{fill description}

\item{ploidy}{fill description}
}
\description{
Internal function
}
