% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_functions.R
\name{GBLUP_genomic_prediction}
\alias{GBLUP_genomic_prediction}
\title{Performes GBLUP}
\usage{
GBLUP_genomic_prediction(
  pheno_dat,
  Geno.mat,
  cycles,
  folds,
  traits,
  cores,
  fixed_cov = NULL,
  fixed_cat = NULL,
  session = NULL
)
}
\arguments{
\item{fixed_cat}{categorical fixed effect}
}
\description{
Performes GBLUP
}
