% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_functions.R
\name{get_relationship_mat}
\alias{get_relationship_mat}
\title{Compute relationship matrix}
\usage{
get_relationship_mat(
  geno_input,
  ped_file,
  type = c("Gmatrix", "Amatrix", "Hmatrix"),
  ploidy,
  pheno
)
}
\arguments{
\item{geno_input}{matrix with individuals in the row and markers in the columns}

\item{ped_file}{pedigree file}

\item{type}{character defining which type}

\item{ploidy}{numeric indicating species ploidy}

\item{pheno}{ToDo}
}
\description{
Compute relationship matrix
}
