% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmatrix2vcf}
\alias{gmatrix2vcf}
\title{Internal function}
\usage{
gmatrix2vcf(Gmat.file, ploidy, output.file, dosageCount = "Reference")
}
\arguments{
\item{Gmat.file}{Genotype matrix with numeric dosage values in the format of samples as columns and SNPs as rows}

\item{ploidy}{species ploidy}

\item{output.file}{path to save the VCF file}

\item{dosageCount}{The format of the dosage values. Default is the count of reference alleles, where 0 = homozygous alternate}
}
\description{
Internal function
}
