% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_functions.R
\name{rrBLUP_genomic_prediction}
\alias{rrBLUP_genomic_prediction}
\title{Function to perform genomic prediction}
\usage{
rrBLUP_genomic_prediction(
  geno,
  pheno,
  traits,
  fixed_effects = NULL,
  fixed_cat = NULL,
  folds = 5,
  iters = 5,
  cores = 1,
  session
)
}
\arguments{
\item{geno}{ToDo}

\item{pheno}{ToDo}

\item{traits}{ToDo}

\item{fixed_effects}{ToDo}

\item{fixed_cat}{categorial fixed effects}

\item{folds}{ToDo}

\item{iters}{ToDo}

\item{cores}{ToDo}

\item{session}{ToDo}
}
\description{
Function to perform genomic prediction
}
