% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_vcf}
\alias{subset_vcf}
\title{Internal function}
\usage{
subset_vcf(vcfR.object, remove.sample.list = NULL, remove.sample.file = NULL)
}
\arguments{
\item{vcfR.object}{vcfR object after importing to R with vcfR::read.vcfR}

\item{remove.sample.list}{A list of sample names to be removed}

\item{remove.sample.file}{(optional) The path to a txt file with a list of sample names to be removed, where each sample is on a new line}
}
\description{
Internal function
}
