% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFun_BIC_Meng.R
\name{CalcBIC}
\alias{CalcBIC}
\title{function for BIC calculation}
\usage{
CalcBIC(y, PC, K)
}
\arguments{
\item{y}{describe documentation}

\item{PC}{describe documentation}

\item{K}{describe documentation}
}
\description{
function for BIC calculation
}
