% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosageCall_functions.R
\name{polyRAD2vcf}
\alias{polyRAD2vcf}
\title{Function to convert polyRAD output to VCF}
\usage{
polyRAD2vcf(geno, model, vcf_path, hindhe.obj, ploidy, output.file, session)
}
\arguments{
\item{geno}{ToDo}

\item{model}{ToDo}

\item{vcf_path}{ToDo}

\item{hindhe.obj}{ToDo}

\item{ploidy}{ToDo}

\item{output.file}{ToDo}

\item{session}{ToDo}
}
\description{
Function to convert polyRAD output to VCF
}
