% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DosageCall_functions.R
\name{polyRAD_dosage_call}
\alias{polyRAD_dosage_call}
\title{Function to perform dosage calling with polyRAD}
\usage{
polyRAD_dosage_call(
  vcf,
  ploidy,
  model,
  p1 = NULL,
  p2 = NULL,
  backcross.gen = 0,
  intercross.gen = 0,
  selfing.gen = 0,
  contamRate = 0.001,
  session
)
}
\arguments{
\item{vcf}{ToDo}

\item{ploidy}{ToDo}

\item{model}{ToDo}

\item{p1}{ToDo}

\item{p2}{ToDo}

\item{backcross.gen}{ToDo}

\item{intercross.gen}{ToDo}

\item{selfing.gen}{ToDo}

\item{contamRate}{ToDo}

\item{session}{ToDo}
}
\description{
Function to perform dosage calling with polyRAD
}
