% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_geno_file}
\alias{read_geno_file}
\title{Read geno file}
\usage{
read_geno_file(file_path, requires = c("GT"), ploidy, check = TRUE)
}
\arguments{
\item{file_path}{character indicanting path to file}

\item{requires}{which information is required from the VCF. Define the FORMAT or INFO letters. Example: c("GT", "DP", "PL")}

\item{ploidy}{numeric value indicating the ploidy of the species}

\item{check}{logical indicating whether to perform checks on the VCF file}
}
\description{
Read geno file
}
