% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcf_sanity_messages}
\alias{vcf_sanity_messages}
\title{Generate Sanity Check Messages for VCF Files}
\usage{
vcf_sanity_messages(checks, required_true, required_false, input_ploidy = NULL)
}
\arguments{
\item{checks}{A \code{vcf_sanity_check} object containing the results of the sanity check.}

\item{required_true}{A character vector of checks that must be TRUE for the VCF file to pass validation.}

\item{required_false}{A character vector of checks that must be FALSE for the VCF file to pass validation.}

\item{input_ploidy}{(Optional) An integer specifying the expected ploidy of the VCF file. If provided, the function will compare it with the detected ploidy.}
}
\value{
None. The function displays alerts for any detected issues.
}
\description{
This function generates user-friendly messages based on the results of a VCF sanity check.
}
\details{
This function uses the results of a VCF sanity check to display appropriate messages to the user. It checks for required conditions that must be met (TRUE or FALSE) and optionally validates the ploidy of the VCF file. If any issues are detected, the function displays alerts using \code{shinyalert}.
}
